+++ hackles/queries/groups/__init__.py
@@ -4,6 +4,8 @@ from .backup_operators_members import get_backup_operators_members
 from .dnsadmins_members import get_dnsadmins_members
 from .gpo_creator_owners import get_gpo_creator_owners
 from .print_operators_members import get_print_operators_members
+from .rodc_allowed_replication import get_rodc_allowed_replication
+from .rodc_denied_missing import get_rodc_denied_missing
 from .schema_admins_nonDA import get_schema_admins_nonDA
 from .server_operators_members import get_server_operators_members
 from .protected_users_missing import get_protected_users_missing
@@ -14,6 +16,8 @@ __all__ = [
     'get_dnsadmins_members',
     'get_gpo_creator_owners',
     'get_print_operators_members',
+    'get_rodc_allowed_replication',
+    'get_rodc_denied_missing',
     'get_schema_admins_nonDA',
     'get_server_operators_members',
     'get_protected_users_missing',