+++ hackles/queries/hygiene/__init__.py
@@ -1,4 +1,5 @@
 """Query functions for hygiene"""
+from .adminsdholder_controllers import get_adminsdholder_controllers
 from .adminsdholder_protected import get_adminsdholder_protected
 from .computer_stale_passwords import get_computer_stale_passwords
 from .computers_without_laps import get_computers_without_laps
@@ -6,6 +7,7 @@ from .enabled_guest_accounts import get_enabled_guest_accounts
 from .krbtgt_age import get_krbtgt_age
 from .ldap_channel_binding import get_ldap_channel_binding
 from .ldap_signing_disabled import get_ldap_signing_disabled
+from .logon_scripts_foreign import get_logon_scripts_foreign
 from .precreated_computers import get_precreated_computers
 from .service_accounts_unprotected import get_service_accounts_unprotected
 from .smb_signing_disabled import get_smb_signing_disabled
@@ -18,6 +20,7 @@ from .users_path_to_da import get_users_path_to_da
 from .privileged_ou_delegation import get_privileged_ou_delegation
 
 __all__ = [
+    'get_adminsdholder_controllers',
     'get_adminsdholder_protected',
     'get_computer_stale_passwords',
     'get_computers_without_laps',
@@ -25,6 +28,7 @@ __all__ = [
     'get_krbtgt_age',
     'get_ldap_channel_binding',
     'get_ldap_signing_disabled',
+    'get_logon_scripts_foreign',
     'get_precreated_computers',
     'get_service_accounts_unprotected',
     'get_smb_signing_disabled',